[ANDDOX-11145] refactored from DoxDialerCaller object to interface with default implementation, per PR suggestion.
