+++ CallWithDoxDialerLib/src/main/java/com/doximity/callwithdoxdialerlib/DoxDialerCaller.kt
@@ -1,14 +1,13 @@
 package com.doximity.callwithdoxdialerlib
 
 import android.content.Context
-import android.content.Intent
-import androidx.core.net.toUri
 
 /**
+ * Interface for making calls through Doximity Dialer.
  *
+ * Use [DoxDialerCaller.getInstance] to get the default implementation.
  */
-object DoxDialerCaller {
-    private const val DOXIMITY_PACKAGE_NAME = "com.doximity.doximitydroid"
+interface DoxDialerCaller {
 
     /**
      * Launches the Doximity app with the specified phone number prefilled in the Dialer dialpad.
@@ -22,11 +21,7 @@ object DoxDialerCaller {
      * with a leading international area code: +1(650)233-3444
      * @return true if Doximity Dialer app is successfully launched or Play Store link is successfully launched, otherwise false.
      */
-    fun dialPhoneNumber(context: Context, phoneNumber: String) = makeCall(
-            context,
-            phoneNumber,
-            CallType.VOICE_PREFILL,
-        )
+    fun dialPhoneNumber(context: Context, phoneNumber: String): Boolean
 
     /**
      * Launches the Doximity app and immediately starts a voice call with the specified phone number.
@@ -39,11 +34,7 @@ object DoxDialerCaller {
      * with a leading international area code: +1(650)233-3444
      * @return true if Doximity Dialer app is successfully launched or Play Store link is successfully launched, otherwise false.
      */
-    fun startVoiceCall(context: Context, phoneNumber: String) = makeCall(
-        context,
-        phoneNumber,
-        CallType.VOICE_DIAL,
-    )
+    fun startVoiceCall(context: Context, phoneNumber: String): Boolean
 
     /**
      * Launches the Doximity app and immediately starts a video call with the specified phone number.
@@ -56,73 +47,17 @@ object DoxDialerCaller {
      * with a leading international area code: +1(650)233-3444
      * @return true if Doximity Dialer app is successfully launched or Play Store link is successfully launched, otherwise false.
      */
-    fun startVideoCall(context: Context, phoneNumber: String) = makeCall(
-        context,
-        phoneNumber,
-        CallType.VIDEO_DIAL,
-    )
+    fun startVideoCall(context: Context, phoneNumber: String): Boolean
 
-    private fun makeCall(
-        context: Context,
-        phoneNumber: String,
-        type: CallType,
-    ): Boolean {
-        return context.packageManager.getLaunchIntentForPackage(DOXIMITY_PACKAGE_NAME)?.apply {
-            addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+    companion object {
+        private val doxDialerCaller: DoxDialerCaller by lazy { DoxDialerCallerImpl() }
 
-            when (type) {
-                CallType.VOICE_PREFILL -> {
-                    setData("https://www.doximity.com/dialer/call?target_number=$phoneNumber".toUri())
-                }
-
-                CallType.VOICE_DIAL -> {
-                    setData("https://www.doximity.com/dialer/call/voice?target_number=$phoneNumber".toUri())
-                }
-
-                CallType.VIDEO_DIAL -> {
-                    setData("https://www.doximity.com/dialer/call/video?target_number=$phoneNumber".toUri())
-                }
-            }
-        }?.let { intent ->
-            try {
-                context.startActivity(intent)
-                true
-            } catch (_: Exception) {
-                false
-            }
-        } ?: run {
-            launchPlayStoreToGetDoxApp(context)
-            true
-        }
-    }
-
-    private fun launchPlayStoreToGetDoxApp(context: Context) {
-        val playStoreIntent = Intent(
-            Intent.ACTION_VIEW,
-            "market://details?id=$DOXIMITY_PACKAGE_NAME".toUri()
-        )
-        val playStoreInstalled = playStoreIntent.resolveActivity(context.packageManager) != null
-        if (playStoreInstalled) {
-            //Open from Singular url for marketing purposes
-            val packageName = context.packageName
-            val singularUrl = "https://doximity.sng.link/Az2j3/c17w?_dl=https://www.doximity.com/dialer/home&_smtype=3&pid=third_party_app&c=$packageName"
-            context.startActivity(
-                Intent(Intent.ACTION_VIEW, singularUrl.toUri())
-            )
-        } else {
-            //Play Store is not installed on user's device, open as a web link
-            context.startActivity(
-                Intent(
-                    Intent.ACTION_VIEW,
-                    "https://play.google.com/store/apps/details?id=$DOXIMITY_PACKAGE_NAME".toUri()
-                )
-            )
-        }
-    }
-
-    private enum class CallType {
-        VOICE_PREFILL,
-        VOICE_DIAL,
-        VIDEO_DIAL,
+        /**
+         * Gets the default implementation of [DoxDialerCaller].
+         *
+         * @return The singleton instance of [DoxDialerCaller]
+         */
+        @JvmStatic
+        fun getInstance(): DoxDialerCaller = doxDialerCaller
     }
 }
